restructuring the code by creating an Utils folder that contains the extractions methods and image processing methods
